fix int overflow in coordinates

+++ src/v-plotter-avr/v-plotter-avr/v.h
@@ -10,13 +10,13 @@
 
 class vPlotter{
 public:
-	unsigned long stepDistance = (unsigned int)(DRIVE_STEP_DISTANCE * 1000); //[um]
-	int X_0;
-	int Y_0;
-	int X_SIZE;
-	int Y_SIZE;
+	unsigned long stepDistance = (unsigned long)(DRIVE_STEP_DISTANCE * 1000); //[um]
+	long X_0;
+	long Y_0;
+	long X_SIZE;
+	long Y_SIZE;
 
-	void setSize(unsigned int xBase) {
+	void setSize(unsigned long xBase) {
 		//automatic calculated printing area for best resolution
 		X_SIZE = int(xBase * 0.4);   //X size of drawing area [mm]
 		X_0 = int(xBase * 0.3); //X offset from drive L [mm]
@@ -30,27 +30,27 @@ public:
         	exit
 		 */
 	}
-	void setSize(unsigned int xBase, int x0, int y0) {
+	void setSize(unsigned long xBase, long x0, long y0) {
 		X_SIZE = xBase - 2 * x0;
 		X_0 = x0;
 		Y_0 = y0;
 		Y_SIZE = X_SIZE;
 	}
 
-	unsigned long getLl(int x, int y) {
+	unsigned long getLl(long x, long y) {
 
 		return sqrt(pow(x + X_0, 2) + pow(y + Y_0, 2));
 	}
 
-	unsigned long getRl(int x, int y) {
+	unsigned long getRl(long x, long y) {
 		return sqrt(pow(X_0 + X_SIZE - x, 2) + pow(Y_0 + y, 2));
 	}
 
-	int getX(unsigned long Ll, unsigned long Rl) {
+	long getX(unsigned long Ll, unsigned long Rl) {
 		return ((pow(Rl, 2) -  pow(Ll, 2) - pow(X_SIZE, 2)) - (2 * X_SIZE * X_0 )) / ((-2 * X_SIZE) - (4 * X_0));
 	}
 
-	int getY(unsigned long Ll, unsigned long Rl) {
+	long getY(unsigned long Ll, unsigned long Rl) {
 		double l2 = pow(Ll, 2) - pow(getX(Ll, Rl) + X_0, 2);
 		if(l2 <= 0)
         	return -1;
